requirements.txt
crewai
crewai_tools
google-generativeai
python-dotenv
embedchain
load_dotenv
embedchain[google]